[WNC-8] refactor code for repository , service of login, register. refactor code for middleware
